allow custom reveal.js themes
