+++ pages/_document.tsx
@@ -1,13 +1,7 @@
 import React from 'react';
 import Document, { Html, Main, NextScript, Head } from 'next/document';
-import { revealCodeThemes, revealThemes } from '../src/components/cdn_assets';
 
 const defaultStyleSheets = [
-  <link
-    key="font-roboto"
-    rel="stylesheet"
-    href="https://fonts.googleapis.com/css?family=Roboto:300,400,500,700&display=swap"
-  />,
   // <link key="font-material-icons" rel="stylesheet" href="https://fonts.googleapis.com/icon?family=Material+Icons" />,
 ] as const;
 
@@ -15,11 +9,12 @@ export default function CustomDocument(): React.ReactElement {
   return (
     <Html>
       <Head>
-        {defaultStyleSheets}
-        {revealThemes.reset}
-        {revealThemes.base}
-        {revealThemes.simple}
-        {revealCodeThemes.docco}
+        <link
+          key="font-roboto"
+          rel="stylesheet"
+          href="https://fonts.googleapis.com/css?family=Roboto:300,400,500,700&display=swap"
+          crossOrigin="anonymous"
+        />
       </Head>
       <body>
         <Main />
